requirements.txt
apache-airflow-providers-amazon
pandas
scikit-learn
python-dotenv
boto3
sagemaker
mlflow>=2.10.0
protobuf
cryptography